refactor auto approval menu to modal
