+++ webview-ui/src/components/chat/auto-approve-menu/types.ts
@@ -1,0 +1,12 @@
+import { AutoApprovalSettings } from "@shared/AutoApprovalSettings"
+
+export interface ActionMetadata {
+	id: keyof AutoApprovalSettings["actions"] | "enableNotifications" | "enableAll"
+	label: string
+	shortName: string
+	description: string
+	icon: string
+	subAction?: ActionMetadata
+	sub?: boolean
+	parentActionId?: string
+}
